+++ Sources/StreamChat/APIClient/APIClient.swift
@@ -6,6 +6,11 @@ import Foundation
 
 /// An object allowing making request to Stream Chat servers.
 class APIClient {
+    private struct RequestsQueueItem {
+        let requestAction: () -> Void
+        let failureAction: () -> Void
+    }
+    
     /// The URL session used for all requests.
     var session: URLSession
     
@@ -15,8 +20,26 @@ class APIClient {
     /// `APIClient` uses this object to decode the results of network requests.
     let decoder: RequestDecoder
     
-    let cdnClient: CDNClient
+    /// Used for reobtaining tokens when they expire and API client receives token expiration error
+    let tokenRefresher: (ClientError, @escaping () -> Void) -> Void
     
+    let cdnClient: CDNClient
+
+    /// Used for synchronizing access to requestsQueue
+    private let requestsAccessQueue = DispatchQueue(label: "io.getstream.requests")
+    
+    /// Stores request failed with token expired error for retrying them later
+    private var requestsQueue = [RequestsQueueItem]()
+    
+    /// Shows whether the token is being refreshed at the moment
+    @Atomic private var isRefreshingToken: Bool = false
+    
+    /// How many times refreshing the token failed consecutively
+    @Atomic private var tokenRefreshConsecutiveFailures: Int = 0
+
+    /// How many times can the token refresh fail before giving up with an error
+    let maxTokenRefreshAttempts = 10
+
     /// Creates a new `APIClient`.
     ///
     /// - Parameters:
@@ -27,12 +50,18 @@ class APIClient {
         sessionConfiguration: URLSessionConfiguration,
         requestEncoder: RequestEncoder,
         requestDecoder: RequestDecoder,
-        CDNClient: CDNClient
+        CDNClient: CDNClient,
+        tokenRefresher: @escaping (ClientError, @escaping () -> Void) -> Void
     ) {
         encoder = requestEncoder
         decoder = requestDecoder
         session = URLSession(configuration: sessionConfiguration)
         cdnClient = CDNClient
+        self.tokenRefresher = tokenRefresher
+    }
+    
+    deinit {
+        requestsQueue = []
     }
     
     /// Performs a network request.
@@ -40,7 +69,19 @@ class APIClient {
     /// - Parameters:
     ///   - endpoint: The `Endpoint` used to create the network request.
     ///   - completion: Called when the networking request is finished.
-    func request<Response: Decodable>(endpoint: Endpoint<Response>, completion: @escaping (Result<Response, Error>) -> Void) {
+    func request<Response: Decodable>(
+        endpoint: Endpoint<Response>,
+        timeout: TimeInterval = 60,
+        completion: @escaping (Result<Response, Error>) -> Void
+    ) {
+        if tokenRefreshConsecutiveFailures > maxTokenRefreshAttempts {
+            return completion(.failure(ClientError.TooManyTokenRefreshAttempts()))
+        }
+
+        if isRefreshingToken {
+            return requeueRequestOnTokenExpired(endpoint: endpoint, timeout: timeout, completion: completion)
+        }
+
         encoder.encodeRequest(for: endpoint) { [weak self] (requestResult) in
             let urlRequest: URLRequest
             do {
@@ -50,7 +91,7 @@ class APIClient {
                 completion(.failure(error))
                 return
             }
-            
+
             log.debug(
                 "Making URL request: \(endpoint.method.rawValue.uppercased()) \(endpoint.path)\n"
                     + "Body:\n\(urlRequest.httpBody?.debugPrettyPrintedJSON ?? "<Empty>")\n"
@@ -61,7 +102,6 @@ class APIClient {
                 log.warning("Callback called while self is nil")
                 return
             }
-            
             let task = self.session.dataTask(with: urlRequest) { [decoder = self.decoder] (data, response, error) in
                 do {
                     let decodedResponse: Response = try decoder.decodeRequestResponse(
@@ -69,12 +109,16 @@ class APIClient {
                         response: response,
                         error: error
                     )
+                    self.tokenRefreshConsecutiveFailures = 0
                     completion(.success(decodedResponse))
                 } catch {
-                    completion(.failure(error))
+                    if error is ClientError.ExpiredToken {
+                        self.requeueRequestOnTokenExpired(endpoint: endpoint, timeout: timeout, completion: completion)
+                    } else {
+                        completion(.failure(error))
+                    }
                 }
             }
-            
             task.resume()
         }
     }
@@ -89,6 +133,52 @@ class APIClient {
             progress: progress,
             completion: completion
         )
+    }
+    
+    /// Queues a failed request for executing later or failing after a timeout
+    func requeueRequestOnTokenExpired<Response: Decodable>(
+        endpoint: Endpoint<Response>,
+        timeout: TimeInterval,
+        completion: @escaping (Result<Response, Error>) -> Void
+    ) {
+        requestsAccessQueue.async {
+            let item = RequestsQueueItem(
+                requestAction: { [weak self] in
+                    self?.request(
+                        endpoint: endpoint,
+                        timeout: timeout,
+                        completion: completion
+                    )
+                },
+                failureAction: { completion(.failure(ClientError.ExpiredToken())) }
+            )
+            
+            self.requestsQueue.append(item)
+            
+            // Only initiate token refreshing once
+            guard self._isRefreshingToken.compareAndSwap(old: false, new: true) else { return }
+            
+            /// Increase the amount of consecutive failures
+            self._tokenRefreshConsecutiveFailures.mutate { $0 += 1 }
+
+            // The moment we queue a first request for execution later we also set a timeout for 60 seconds
+            // after which we fail all the queued requests
+            self.requestsAccessQueue.asyncAfter(deadline: .now() + timeout) {
+                self.requestsQueue.forEach { $0.failureAction() }
+                self.requestsQueue = []
+            }
+            
+            self.tokenRefresher(ClientError.ExpiredToken()) { [weak self] in
+                guard let self = self else {
+                    return
+                }
+
+                self.isRefreshingToken = false
+                let queue = self.requestsQueue
+                self.requestsQueue = []
+                queue.forEach { $0.requestAction() }
+            }
+        }
     }
 }