+++ Sources/StreamChat/ChatClient.swift
@@ -64,7 +64,9 @@ public class ChatClient {
     }()
     
     /// The `APIClient` instance `Client` uses to communicate with Stream REST API.
-    lazy var apiClient: APIClient = {
+    lazy var apiClient: APIClient = makeAPIClient()
+        
+    private func makeAPIClient() -> APIClient {
         var encoder = environment.requestEncoderBuilder(config.baseURL.restAPIBaseURL, config.apiKey)
         encoder.connectionDetailsProviderDelegate = self
         
@@ -78,10 +80,17 @@ public class ChatClient {
                 encoder: encoder,
                 decoder: decoder,
                 sessionConfiguration: urlSessionConfiguration
-            )
+            ),
+            { [weak self] error, completion in
+                guard let self = self else {
+                    completion()
+                    return
+                }
+                self.refreshToken(error: error, completion: { _ in completion() })
+            }
         )
         return apiClient
-    }()
+    }
     
     /// The `WebSocketClient` instance `Client` uses to communicate with Stream WS servers.
     lazy var webSocketClient: WebSocketClient? = {
@@ -338,6 +347,7 @@ public class ChatClient {
         clientUpdater.disconnect()
         userConnectionProvider = nil
         unsubscribeFromNotifications()
+        apiClient = makeAPIClient()
     }
 
     func fetchCurrentUserIdFromDatabase() -> UserId? {
@@ -485,13 +495,15 @@ extension ChatClient {
             _ sessionConfiguration: URLSessionConfiguration,
             _ requestEncoder: RequestEncoder,
             _ requestDecoder: RequestDecoder,
-            _ CDNClient: CDNClient
+            _ CDNClient: CDNClient,
+            _ tokenRefresher: @escaping (ClientError, @escaping () -> Void) -> Void
         ) -> APIClient = {
             APIClient(
                 sessionConfiguration: $0,
                 requestEncoder: $1,
                 requestDecoder: $2,
-                CDNClient: $3
+                CDNClient: $3,
+                tokenRefresher: $4
             )
         }
         
@@ -607,7 +619,7 @@ extension ChatClient: ConnectionStateDelegate {
         case let .disconnected(error: error):
             if let error = error,
                error.isTokenExpiredError {
-                refreshToken(error: error)
+                refreshToken(error: error, completion: nil)
                 shouldNotifyConnectionIdWaiters = false
             } else {
                 shouldNotifyConnectionIdWaiters = true
@@ -627,7 +639,10 @@ extension ChatClient: ConnectionStateDelegate {
         )
     }
     
-    private func refreshToken(error: ClientError) {
+    private func refreshToken(
+        error: ClientError,
+        completion: ((Error?) -> Void)?
+    ) {
         guard let tokenProvider = tokenProvider else {
             return log.assertionFailure(
                 "In case if token expiration is enabled on backend you need to provide a way to reobtain it via `tokenProvider` on ChatClient"
@@ -646,14 +661,13 @@ extension ChatClient: ConnectionStateDelegate {
                 clientUpdater.reloadUserIfNeeded(
                     userConnectionProvider: .closure { _, completion in
                         tokenProvider() { result in
-                            if case let .success(token) = result,
-                               !token.isExpired {
+                            if case .success = result {
                                 self.tokenExpirationRetryStrategy.successfullyConnected()
                             }
-                            
                             completion(result)
                         }
-                    }
+                    },
+                    completion: completion
                 )
             }
     }