+++ CHANGELOG.md
@@ -3,6 +3,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🐞 Fixed
+- Fix token expiration refresh mechanism for API endpoints
+
 ### 🔄 Changed
 
 # [4.0.1](https://github.com/GetStream/stream-chat-swift/releases/tag/4.0.1)